Device dashboard values refactor

+++ src/integrations/iron_nest/types/mod.rs
@@ -38,7 +38,7 @@ pub struct Device {
     pub name: String,
     pub device_type: DeviceType,
     pub ip: String,
-    pub power_state: i64,
+    pub power_state: i32,
     pub battery_percentage: i64,
     pub last_seen: DateTime<Utc>,
 }
